Fix variant availablity test
