+++ saleor/graphql/product/tests/test_variant_availability.py
@@ -6,77 +6,19 @@ from ...tests.utils import get_graphql_content
 COUNTRY_CODE = "US"
 
 
-QUERY_DEPRECATED_VARIANT_AVAILABILITY = """
-    query variantAvailability($id: ID!) {
-        productVariant(id: $id) {
-            isAvailable
-            stockQuantity
-            quantityAvailable
-        }
-    }
-"""
-
-
-def test_variant_availability_without_inventory_tracking(
-    api_client, variant_without_inventory_tracking, settings
-):
-    variant = variant_without_inventory_tracking
-    variables = {"id": graphene.Node.to_global_id("ProductVariant", variant.pk)}
-    response = api_client.post_graphql(QUERY_DEPRECATED_VARIANT_AVAILABILITY, variables)
-    content = get_graphql_content(response)
-    variant_data = content["data"]["productVariant"]
-    assert variant_data["isAvailable"] is True
-    assert variant_data["stockQuantity"] == settings.MAX_CHECKOUT_LINE_QUANTITY
-    assert variant_data["quantityAvailable"] == settings.MAX_CHECKOUT_LINE_QUANTITY
-
-
-def test_variant_availability(api_client, variant_with_many_stocks, settings):
-    variant = variant_with_many_stocks
-    variables = {"id": graphene.Node.to_global_id("ProductVariant", variant.pk)}
-    response = api_client.post_graphql(QUERY_DEPRECATED_VARIANT_AVAILABILITY, variables)
-    content = get_graphql_content(response)
-    variant_data = content["data"]["productVariant"]
-    assert variant_data["isAvailable"] is True
-    assert variant_data["stockQuantity"] == 7
-    assert variant_data["quantityAvailable"] == 7
-
-
-def test_variant_availability_without_inventory_tracking_without_stocks(
-    api_client, variant_without_inventory_tracking, settings
-):
-    variant = variant_without_inventory_tracking
-    variant.stocks.all().delete()
-    variables = {"id": graphene.Node.to_global_id("ProductVariant", variant.pk)}
-    response = api_client.post_graphql(QUERY_DEPRECATED_VARIANT_AVAILABILITY, variables)
-    content = get_graphql_content(response)
-    variant_data = content["data"]["productVariant"]
-    assert variant_data["isAvailable"] is True
-    assert variant_data["stockQuantity"] == settings.MAX_CHECKOUT_LINE_QUANTITY
-    assert variant_data["quantityAvailable"] == settings.MAX_CHECKOUT_LINE_QUANTITY
-
-
-def test_variant_availability_without_stocks(api_client, variant, settings):
-    variables = {"id": graphene.Node.to_global_id("ProductVariant", variant.pk)}
-    response = api_client.post_graphql(QUERY_DEPRECATED_VARIANT_AVAILABILITY, variables)
-    content = get_graphql_content(response)
-    variant_data = content["data"]["productVariant"]
-    assert variant_data["isAvailable"] is False
-    assert variant_data["stockQuantity"] == 0
-    assert variant_data["quantityAvailable"] == 0
-
-
 def test_variant_quantity_available_without_country_code(
-    api_client, variant_with_many_stocks
+    api_client, variant_with_many_stocks, channel_USD
 ):
     query = """
-    query variantAvailability($id: ID!) {
-        productVariant(id: $id) {
+    query variantAvailability($id: ID!, $channelSlug: String) {
+        productVariant(id: $id, channelSlug: $channelSlug) {
             quantityAvailable
         }
     }
     """
     variables = {
-        "id": graphene.Node.to_global_id("ProductVariant", variant_with_many_stocks.pk)
+        "id": graphene.Node.to_global_id("ProductVariant", variant_with_many_stocks.pk),
+        "channelSlug": channel_USD.slug,
     }
     response = api_client.post_graphql(query, variables)
     content = get_graphql_content(response)
@@ -85,8 +27,8 @@ def test_variant_quantity_available_without_country_code(
 
 
 QUERY_VARIANT_AVAILABILITY = """
-    query variantAvailability($id: ID!, $country: CountryCode) {
-        productVariant(id: $id) {
+    query variantAvailability($id: ID!, $country: CountryCode, $channelSlug: String) {
+        productVariant(id: $id, channelSlug: $channelSlug) {
             quantityAvailable(countryCode: $country)
         }
     }
@@ -94,11 +36,12 @@ QUERY_VARIANT_AVAILABILITY = """
 
 
 def test_variant_quantity_available_with_country_code(
-    api_client, variant_with_many_stocks,
+    api_client, variant_with_many_stocks, channel_USD
 ):
     variables = {
         "id": graphene.Node.to_global_id("ProductVariant", variant_with_many_stocks.pk),
         "country": COUNTRY_CODE,
+        "channelSlug": channel_USD.slug,
     }
     response = api_client.post_graphql(QUERY_VARIANT_AVAILABILITY, variables)
     content = get_graphql_content(response)
@@ -107,11 +50,12 @@ def test_variant_quantity_available_with_country_code(
 
 
 def test_variant_quantity_available_with_null_as_country_code(
-    api_client, variant_with_many_stocks
+    api_client, variant_with_many_stocks, channel_USD
 ):
     variables = {
         "id": graphene.Node.to_global_id("ProductVariant", variant_with_many_stocks.pk),
         "country": None,
+        "channelSlug": channel_USD.slug,
     }
     response = api_client.post_graphql(QUERY_VARIANT_AVAILABILITY, variables)
     content = get_graphql_content(response)
@@ -121,7 +65,7 @@ def test_variant_quantity_available_with_null_as_country_code(
 
 @override_settings(MAX_CHECKOUT_LINE_QUANTITY=15)
 def test_variant_quantity_available_with_max(
-    api_client, variant_with_many_stocks, settings
+    api_client, variant_with_many_stocks, settings, channel_USD
 ):
     stock = variant_with_many_stocks.stocks.first()
     stock.quantity = 16
@@ -129,6 +73,7 @@ def test_variant_quantity_available_with_max(
     variables = {
         "id": graphene.Node.to_global_id("ProductVariant", variant_with_many_stocks.pk),
         "country": COUNTRY_CODE,
+        "channelSlug": channel_USD.slug,
     }
     response = api_client.post_graphql(QUERY_VARIANT_AVAILABILITY, variables)
     content = get_graphql_content(response)
@@ -137,12 +82,13 @@ def test_variant_quantity_available_with_max(
 
 
 def test_variant_quantity_available_without_stocks(
-    api_client, variant_with_many_stocks,
+    api_client, variant_with_many_stocks, channel_USD
 ):
     variant_with_many_stocks.stocks.all().delete()
     variables = {
         "id": graphene.Node.to_global_id("ProductVariant", variant_with_many_stocks.pk),
         "country": COUNTRY_CODE,
+        "channelSlug": channel_USD.slug,
     }
     response = api_client.post_graphql(QUERY_VARIANT_AVAILABILITY, variables)
     content = get_graphql_content(response)
@@ -155,10 +101,12 @@ def test_variant_quantity_available_with_allocations(
     variant_with_many_stocks,
     order_line_with_allocation_in_many_stocks,
     order_line_with_one_allocation,
+    channel_USD,
 ):
     variables = {
         "id": graphene.Node.to_global_id("ProductVariant", variant_with_many_stocks.pk),
         "country": COUNTRY_CODE,
+        "channelSlug": channel_USD.slug,
     }
     response = api_client.post_graphql(QUERY_VARIANT_AVAILABILITY, variables)
     content = get_graphql_content(response)
@@ -168,13 +116,14 @@ def test_variant_quantity_available_with_allocations(
 
 @override_settings(MAX_CHECKOUT_LINE_QUANTITY=15)
 def test_variant_quantity_available_without_inventory_tracking(
-    api_client, variant_with_many_stocks, settings
+    api_client, variant_with_many_stocks, settings, channel_USD
 ):
     variant_with_many_stocks.track_inventory = False
     variant_with_many_stocks.save(update_fields=["track_inventory"])
     variables = {
         "id": graphene.Node.to_global_id("ProductVariant", variant_with_many_stocks.pk),
         "country": COUNTRY_CODE,
+        "channelSlug": channel_USD.slug,
     }
     response = api_client.post_graphql(QUERY_VARIANT_AVAILABILITY, variables)
     content = get_graphql_content(response)
@@ -184,13 +133,14 @@ def test_variant_quantity_available_without_inventory_tracking(
 
 @override_settings(MAX_CHECKOUT_LINE_QUANTITY=15)
 def test_variant_quantity_available_without_inventory_tracking_and_stocks(
-    api_client, variant, settings
+    api_client, variant, settings, channel_USD
 ):
     variant.track_inventory = False
     variant.save(update_fields=["track_inventory"])
     variables = {
         "id": graphene.Node.to_global_id("ProductVariant", variant.pk),
         "country": COUNTRY_CODE,
+        "channelSlug": channel_USD.slug,
     }
     response = api_client.post_graphql(QUERY_VARIANT_AVAILABILITY, variables)
     content = get_graphql_content(response)
